Changed filename, now returns a list of Killmail obj.
